+++ src/c++/uds/src/uds/uds-threads.h
@@ -32,16 +32,6 @@ struct cond_var {
 
 struct thread;
 
-struct barrier {
-	/* Lock for this barrier object */
-	struct semaphore lock;
-	/* Semaphore for threads waiting at the barrier */
-	struct semaphore wait;
-	/* Number of threads which have arrived */
-	int arrived;
-	/* Total number of threads using this barrier */
-	int thread_count;
-};
 #else
 struct cond_var {
 	pthread_cond_t condition;
@@ -81,11 +71,6 @@ int __must_check uds_create_thread(void (*thread_function)(void *), void *thread
 void uds_perform_once(atomic_t *once_state, void (*function) (void));
 
 int uds_join_threads(struct thread *thread);
-
-int __must_check uds_initialize_barrier(struct barrier *barrier,
-					unsigned int thread_count);
-int uds_destroy_barrier(struct barrier *barrier);
-int uds_enter_barrier(struct barrier *barrier);
 
 int __must_check uds_init_cond(struct cond_var *cond);
 int uds_signal_cond(struct cond_var *cond);
@@ -195,6 +180,11 @@ int __must_check uds_init_mutex(struct mutex *mutex);
 int uds_destroy_mutex(struct mutex *mutex);
 void uds_lock_mutex(struct mutex *mutex);
 void uds_unlock_mutex(struct mutex *mutex);
+
+int __must_check uds_initialize_barrier(struct barrier *barrier,
+					unsigned int thread_count);
+int uds_destroy_barrier(struct barrier *barrier);
+int uds_enter_barrier(struct barrier *barrier);
 
 int __must_check uds_initialize_semaphore(struct semaphore *semaphore,
 					  unsigned int value);
